requirements.txt
# Core Prefect
prefect>=3.0.0
prefect-client>=3.0.0

# AI/ML
anthropic>=0.39.0

# Integrations
slack-sdk>=3.33.0

# Observability
logfire>=0.70.0

# MCP (Model Context Protocol)
# Note: Update with actual MCP client package when available
# For now, assuming HTTP client for MCP server
httpx>=0.27.0

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-mock>=3.14.0
pytest-cov>=5.0.0

# Development
black>=24.0.0
ruff>=0.6.0
mypy>=1.11.0
